UI: Base for claims updated component.
