+++ src/components/play-mode/claims/claims.module.css
@@ -1,0 +1,91 @@
+
+.claimsWrapper {
+    position: absolute;
+    bottom: 135px;
+    left: 325px;
+    min-width: 115px;
+    display: none;
+}
+
+.claimsWrapper.show {
+    display: block;
+}
+
+.claimsBtn {
+    display: inline-block;
+    position: absolute;
+    top: 0px;
+    left: 0px;
+    padding: 10px;
+    background-color: #000;
+    font-size: 16px;
+    color: #FFF;
+    cursor: pointer;
+    user-select: none;
+    overflow: hidden;
+    transition: color 1s cubic-bezier(0, 1, 0, 1);
+    transform: skew(-20deg) translate(-10px, -10px);
+    z-index: 1;
+}
+
+.claimsBtn .background {
+    position: absolute;
+    top: 0;
+    bottom: 0;
+    height: 100%;
+    transition: transform 1s cubic-bezier(0, 1, 0, 1);
+    left: 0;
+    transform: rotateZ(180deg) translateX(-100%);
+    filter: invert(50%) sepia(99%) saturate(514%) hue-rotate(177deg) brightness(99%) contrast(94%);
+}
+
+.claimsBtn .text {
+    position: relative;
+    font-size: 16px;
+    color: #FFF;
+    cursor: pointer;
+    user-select: none;
+}
+
+/*  */
+
+.claimsPanel {
+    position: relative;
+    width: 0px;
+    height: 0px;
+    background-color: #000;
+    opacity: 0;
+    transition: width 300ms ease, height 300ms ease, opacity 300ms ease;
+    overflow: hidden;
+}
+
+.claimsPanel.opened {
+    width: 300px;
+    height: 200px;
+    opacity: 1;
+    overflow-y: scroll;
+}
+
+.buttonsBar {
+    position: absolute;
+    top: 0px;
+    right: 0px;
+}
+
+/*  */
+
+.claimsList {
+    position: absolute;
+    top: 40px;
+    bottom: 0px;
+    left: 0px;
+    width: 100%;
+}
+
+.claimItem {
+    width: 100%;
+    height: 20px;
+    color: #fff;
+    text-align: left;
+    padding-left: 20px;
+}
