+++ src/components/play-mode/claims/index.tsx
@@ -1,0 +1,3 @@
+
+import { Claims } from './Claims';
+export { Claims };
